improving code coverage
